diktatFix
